Ambiente de teste para usuarios terminado

+++ database/migrations/2018_01_21_115649_create_latitudes_table.php
@@ -14,7 +14,7 @@ class CreateLatitudesTable extends Migration
     public function up()
     {
         Schema::create('latitudes', function (Blueprint $table) {
-					$table->float('usr_lat')->primary();
+					$table->double('usr_lat')->primary();
 					$table->string('usr_name');
 					$table->foreign('usr_name')->references('usr_name')->on('logins')->onDelete('CASCADE');
 					$table->timestamps();
